refactor: Extract PaneState into separate module
Extract the large PaneState struct and implementation (~700 lines) from core.rs
into a dedicated pane_state.rs module to improve code organization and maintainability.

Changes:
- Create new pane_state.rs module with PaneState and related types
- Update core.rs to import PaneState from new module location
- Update mod.rs to include and re-export pane_state module
- Fix toggle_pane method to use explicit match instead of Not operator
- Clean up unused imports in both modules
